+++ cli/src/pcluster/schemas/cluster_schema.py
@@ -125,14 +125,9 @@ from pcluster.constants import (
     SUPPORTED_OSES,
 )
 from pcluster.models.s3_bucket import parse_bucket_url
-from pcluster.schemas.common_schema import (
-    AdditionalIamPolicySchema,
-    BaseDevSettingsSchema,
-    BaseSchema,
-    TagSchema,
-    get_field_validator,
-    validate_no_reserved_tag,
-)
+from pcluster.schemas.common_schema import AdditionalIamPolicySchema, BaseDevSettingsSchema, BaseSchema
+from pcluster.schemas.common_schema import ImdsSchema as TopLevelImdsSchema
+from pcluster.schemas.common_schema import TagSchema, get_field_validator, validate_no_reserved_tag
 from pcluster.utils import yaml_load
 from pcluster.validators.cluster_validators import EFS_MESSAGES, FSX_MESSAGES
 
@@ -1973,6 +1968,7 @@ class ClusterSchema(BaseSchema):
         DirectoryServiceSchema, metadata={"update_policy": UpdatePolicy.COMPUTE_FLEET_STOP}
     )
     config_region = fields.Str(data_key="Region", metadata={"update_policy": UpdatePolicy.UNSUPPORTED})
+    imds = fields.Nested(TopLevelImdsSchema, metadata={"update_policy": UpdatePolicy.UNSUPPORTED})
     custom_s3_bucket = fields.Str(metadata={"update_policy": UpdatePolicy.READ_ONLY_RESOURCE_BUCKET})
     additional_resources = fields.Str(metadata={"update_policy": UpdatePolicy.SUPPORTED})
     dev_settings = fields.Nested(ClusterDevSettingsSchema, metadata={"update_policy": UpdatePolicy.SUPPORTED})